+++ backend/src/main/java/com/carffeine/carffeine/station/infrastructure/repository/review/ReviewQueryRepository.java
@@ -1,5 +1,7 @@
 package com.carffeine.carffeine.station.infrastructure.repository.review;
 
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.ReviewResponse;
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.TotalRatingsResponse;
 import com.querydsl.core.QueryResults;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
