+++ backend/src/main/java/com/carffeine/carffeine/station/infrastructure/repository/review/dto/TotalRatingsResponse.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.station.infrastructure.repository.review;
+package com.carffeine.carffeine.station.infrastructure.repository.review.dto;
 
 public record TotalRatingsResponse(
         Double totalRatings,
