+++ backend/src/test/java/com/carffeine/carffeine/station/controller/review/ReviewControllerTest.java
@@ -4,9 +4,9 @@ import com.carffeine.carffeine.helper.MockBeanInjection;
 import com.carffeine.carffeine.member.domain.Member;
 import com.carffeine.carffeine.station.domain.review.Review;
 import com.carffeine.carffeine.station.domain.station.Station;
-import com.carffeine.carffeine.station.infrastructure.repository.review.ReviewResponse;
-import com.carffeine.carffeine.station.infrastructure.repository.review.ReviewResponses;
-import com.carffeine.carffeine.station.infrastructure.repository.review.TotalRatingsResponse;
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.ReviewResponse;
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.ReviewResponses;
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.TotalRatingsResponse;
 import com.carffeine.carffeine.station.service.review.dto.CreateReviewRequest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayNameGeneration;
