+++ backend/src/test/java/com/carffeine/carffeine/station/service/review/ReviewQueryServiceTest.java
@@ -8,9 +8,9 @@ import com.carffeine.carffeine.station.domain.review.ReviewRepository;
 import com.carffeine.carffeine.station.domain.station.Station;
 import com.carffeine.carffeine.station.domain.station.StationRepository;
 import com.carffeine.carffeine.station.infrastructure.repository.review.ReviewQueryRepository;
-import com.carffeine.carffeine.station.infrastructure.repository.review.ReviewResponse;
-import com.carffeine.carffeine.station.infrastructure.repository.review.ReviewResponses;
-import com.carffeine.carffeine.station.infrastructure.repository.review.TotalRatingsResponse;
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.ReviewResponse;
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.ReviewResponses;
+import com.carffeine.carffeine.station.infrastructure.repository.review.dto.TotalRatingsResponse;
 import com.carffeine.carffeine.station.service.review.dto.CreateReviewRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
